feat: separate menubar from topbar
BREAKING CHANGE: topbar no longer renders role=menubar. New component <MenuBar /> must now be used if menubar is desired
