+++ docs/index.js
@@ -6,6 +6,7 @@ import classNames from 'classnames';
 import Home from './Home';
 import {
   TopBar,
+  MenuBar,
   TopBarTrigger,
   TopBarItem,
   Workspace,
@@ -128,56 +129,58 @@ class App extends Component {
             defaultTitle="Deque Cauldron React"
           />
           <SkipLink target={'#main-content'} />
-          <TopBar thin={thin} hasTrigger>
-            <TopBarTrigger onClick={this.onTriggerClick}>
-              <button
-                tabIndex={-1}
-                aria-label="Menu"
-                aria-haspopup="true"
-                ref={this.topBarTrigger}
-                aria-expanded={show}
+          <TopBar>
+            <MenuBar thin={thin} hasTrigger>
+              <TopBarTrigger onClick={this.onTriggerClick}>
+                <button
+                  tabIndex={-1}
+                  aria-label="Menu"
+                  aria-haspopup="true"
+                  ref={this.topBarTrigger}
+                  aria-expanded={show}
+                >
+                  <Icon type="fa-bars" />
+                </button>
+              </TopBarTrigger>
+              <TopBarItem>
+                <Link to="/" className="MenuItem__logo" tabIndex={-1}>
+                  <img src={logo} alt="" /> <span>Cauldron</span>
+                </Link>
+              </TopBarItem>
+
+              {/* The below line demonstrates the ability to conditionally include menu item children. */}
+              {false && <TopBarItem>Potato</TopBarItem>}
+
+              <TopBarMenu
+                id="topbar-menu"
+                className="MenuItem--align-right MenuItem--separator MenuItem--arrow-down"
+                menuItemRef={el => (this.topBarMenuItem = el)}
               >
-                <Icon type="fa-bars" />
-              </button>
-            </TopBarTrigger>
-            <TopBarItem>
-              <Link to="/" className="MenuItem__logo" tabIndex={-1}>
-                <img src={logo} alt="" /> <span>Cauldron</span>
-              </Link>
-            </TopBarItem>
+                <div className="TopBar__item--icon">
+                  {thin ? (
+                    <Icon type="fa-cog" label="Settings" />
+                  ) : (
+                    <Fragment>
+                      <Icon type="fa-cog" />
+                      <div>Settings</div>
+                    </Fragment>
+                  )}
+                </div>
+                <OptionsMenuList onSelect={this.onSettingsSelect}>
+                  <li>Default top bar</li>
+                  <li>Thin top bar</li>
+                </OptionsMenuList>
+              </TopBarMenu>
 
-            {/* The below line demonstrates the ability to conditionally include menu item children. */}
-            {false && <TopBarItem>Potato</TopBarItem>}
-
-            <TopBarMenu
-              id="topbar-menu"
-              className="MenuItem--align-right MenuItem--separator MenuItem--arrow-down"
-              menuItemRef={el => (this.topBarMenuItem = el)}
-            >
-              <div className="TopBar__item--icon">
-                {thin ? (
-                  <Icon type="fa-cog" label="Settings" />
-                ) : (
-                  <Fragment>
-                    <Icon type="fa-cog" />
-                    <div>Settings</div>
-                  </Fragment>
-                )}
-              </div>
-              <OptionsMenuList onSelect={this.onSettingsSelect}>
-                <li>Default top bar</li>
-                <li>Thin top bar</li>
-              </OptionsMenuList>
-            </TopBarMenu>
-
-            <TopBarItem className="MenuItem--separator">
-              <a
-                href="https://github.com/dequelabs/cauldron-react"
-                className="fa fa-github"
-                aria-label="Cauldron React on GitHub"
-                tabIndex={-1}
-              />
-            </TopBarItem>
+              <TopBarItem className="MenuItem--separator">
+                <a
+                  href="https://github.com/dequelabs/cauldron-react"
+                  className="fa fa-github"
+                  aria-label="Cauldron React on GitHub"
+                  tabIndex={-1}
+                />
+              </TopBarItem>
+            </MenuBar>
           </TopBar>
           <SideBar show={this.state.show} onDismiss={this.onTriggerClick}>
             {componentsList.map(name => {
